Remove the use of getPackageIterator in PackageManager.loadSCMPackage
Using `getPackageIterator` means we need to have scanned all locations.
However, we want to move to a position where a simple `dub build` with
resolved dependencies does not trigger a full scan, and this would
prevent it. However, as the comment explain, we don't actually need
to use `getPackageIterator` here and instead can do lazy loading.